pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nba-analytics"
version = "0.1.0"
description = "Standalone analytics pipeline for NBA Daily API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests==2.32.3",
  "psycopg[binary]==3.2.9",
  "python-dotenv==1.0.1",
  "streamlit==1.43.1",
  "pandas==2.2.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
